Data.Attoparsec.Lazy

Description

Simple, efficient combinator parsing for lazy ByteString strings, loosely based on the Parsec library.

This is essentially the same code as in the Attoparsec module, only with a parse function that can consume a lazy ByteString incrementally, and a Result type that does not allow more input to be fed in. Think of this as suitable for use with a lazily read file, e.g. via readFile or hGetContents.

Behind the scenes, strict ByteString values are still used internally to store parser input and manipulate it efficiently. High-performance parsers such as string still expect strict ByteString parameters.

Documentation

module Data.Attoparsec.ByteString.Lazy