Data.Attoparsec.Number

Description

Deprecated: This module will be removed in the next major release.

This module is deprecated, and both the module and Number type will be removed in the next major release. Use the scientific package and the Scientific type instead.

A simple number type, useful for parsing both exact and inexact quantities without losing much precision.

Documentation

data Number

A numeric type that can represent integers accurately, and floating point numbers to the precision of a Double.

Note: this type is deprecated, and will be removed in the next major release. Use the Scientific type instead.

Constructors

I !Integer

D !Double