Data.Attoparsec.Text.Buffer

Description

An immutable buffer that supports cheap appends.

Documentation

data Buffer

buffer :: Text -> Buffer

The initial Buffer has no mutable zone, so we can avoid all copies in the (hopefully) common case of no further input being fed to us.

unbuffer :: Buffer -> Text

unbufferAt :: Int -> Buffer -> Text

length :: Buffer -> Int

pappend :: Buffer -> Text -> Buffer

iter :: Buffer -> Int -> Iter

O(1) Iterate (unsafely) one step forwards through a UTF-8 array, returning the current character and the delta to add to give the next offset to iterate at.

iter_ :: Buffer -> Int -> Int

O(1) Iterate one step through a UTF-8 array, returning the delta to add to give the next offset to iterate at.

substring :: Int -> Int -> Buffer -> Text

lengthCodeUnits :: Text -> Int

dropCodeUnits :: Int -> Buffer -> Text