Data.Attoparsec.Text.FastSet

Description

Fast set membership tests for Char values. We test for membership using a hashtable implemented with Robin Hood collision resolution. The set representation is unboxed, and the characters and hashes interleaved, for efficiency.

Data type

data FastSet

Construction

fromList :: String -> FastSet

set :: Text -> FastSet

Lookup

member :: Char -> FastSet -> Bool

Check the set for membership.

Handy interface

charClass :: String -> FastSet