Data.Attoparsec.Number

Description

A simple number type, useful for parsing both exact and inexact quantities without losing much precision.

Documentation

data Number

A numeric type that can represent integers accurately, and floating point numbers to the precision of a Double.

Constructors

I !Integer

D !Double